New method of checking for BWF signals
This method stores the code that was picked up whether it matches a
known code or not. The reason for this change is twofold.

1: The function readSensor runs every time a pulse is picked up by the
currently selected BWF sensor coil. If another pulse arrives while the
function is still running, it will be discarded. Therefore, it is vital
that the readSensor function is kept as short and fast as possible, and
all heavy analysis should be made at another time.

2: The new matchCode function can match any code, not only the
inside/outside BWF codes. This could - in theory - be used for finding
guide wires.

The functions isTimedOut and hasNoSignal have been removed. Of those
two only isTimedOut was used, but in a way that it's result didn't
matter.

Also, the logic for BWF_DETECTION_ALWAYS has been moved into the BWF
code.